manager/python/geoprocessing/pyproject.toml
[tool.poetry]
name = "braga-task"
version = "0.1.0"
description = ""
authors = ["Miko <jherjati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
dramatiq-pg = "^0.11.0"
psycopg2-binary = "^2.9.8"
python-dotenv = "^1.0.0"
dramatiq = "^1.14.2"
minio = "^7.1.17"
py3dtiles = {extras = ["las"], version = "^7.0.0"}
laspy = {extras = ["laszip", "lazrs"], version = "^2.5.3"}

[tool.poetry.group.dev.dependencies]
watchdog = "^3.0.0"
gdal = "3.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
